+++ packages/admin/src/mui/menu/CollapsibleItem.tsx
@@ -17,6 +17,7 @@ export interface ICollapsibleItemProps extends IMenuItemProps {
     collapsible: boolean;
     secondaryAction?: React.ComponentType<ICollapsibleItemSecondaryActionProps>;
     isOpen?: boolean;
+    textComponent?: React.ReactNode;
 }
 
 export interface ICollapsibleItemSecondaryActionProps {
@@ -34,6 +35,7 @@ export const MenuCollapsibleItem: React.FunctionComponent<ICollapsibleItemProps 
     text,
     icon,
     children,
+    textComponent,
     secondaryAction: SecondaryAction = DefaultSecondaryAction,
     ...otherProps
 }) => {
@@ -51,7 +53,7 @@ export const MenuCollapsibleItem: React.FunctionComponent<ICollapsibleItemProps 
                 {...{ text, icon, level }}
                 onClick={handleClick.bind(null, open, setOpen)}
                 secondaryAction={collapsible && <SecondaryAction open={open} />}
-            />
+            >{textComponent}</MenuItem>
             {collapsible ? (
                 <Collapse in={open} timeout="auto" unmountOnExit>
                     {childElements}