+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+### Internal Changes
+
+-   Pass `children` to `MenuItem` component
+-   Pass `textComponent` to `MenuCollapsibleItem` component
+
 ## [1.3.0] - 4. March 2021
 
 This is a bugfix/maintenance release.